gemini-cli: extract event-translation helpers into events.py
After ``fe43674 feat(providers): add structured stream logging`` landed,
``backend/app/core/providers/gemini_cli/client.py`` ballooned to 551
lines and started failing the 500-line file budget on every PR. The
gate failure was masking real PR check output and forcing every
downstream PR (#438, #440, #441, #442, #443, #444) to inherit a red
``bun run check`` job.

Split the module by concern (no behaviour change):

* ``client.py`` keeps the long-lived stateful ``PawrrtalAcpClient``
  class plus the helpers it consults directly during request handling
  (``pick_allow_option``, ``ALLOW_KINDS``, ``_DENY_OUTCOME`` etc.).
  Now 346 lines.
* ``events.py`` owns the pure format-translation surface — ACP
  ``session/update`` variant → Pawrrtal ``StreamEvent`` (or ``None``
  for the variants we intentionally drop), plus the structured
  operator-log formatter that fires alongside every update. 250 lines.

``client.py`` re-imports the two helpers it still needs
(``_log_session_update``, ``_stream_event_for_update``) so the
session_update hot path stays a one-line dispatch. ``test_gemini_cli_
provider`` now imports the private helpers from ``events`` instead of
``client``; everything else (78 tests across that file) keeps passing.

Re-running the file-budget gate locally confirms no file in the repo
exceeds 500 lines now.
